pyproject.toml
[tool.poetry]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = ["Oleg Smirnov <oleg.smirnoff505@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
django = "4.2"
python-dotenv = "^1.0.1"
ipython = "^8.24.0"
djangorestframework = "^3.15.1"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
